Provide Libtool's `.la` files once again.
I had intended not to include these files in the original package, but due to
a typo, they had been included. As alluded to in the code comment in
`build.sh`, removing these files turns out to be a breaking change that has
been messing up builds. At the moment, the most sensible course of action
seems to be to go back to providing these. However, we start emulating Debian
by blanking out the `dependency_libs` item in the `.la` files, which causes
the interdependencies and is not needed now that virtually everyone uses
`pkg-config`.